Seperates ingestion into alert_acquisition and data_upload modules
